Isolated Azure Storage Functions to a File Storage Service
